drop "isolated" from name
