+++ systemtests/src/test/java/io/enmasse/systemtest/iot/tls/TlsVersionTest.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.systemtest.iot.isolated.tls;
+package io.enmasse.systemtest.iot.tls;
 
 import io.enmasse.systemtest.iot.HttpAdapterClient;
 import io.enmasse.systemtest.iot.IoTTestSession;
