+++ systemtests/src/test/java/io/enmasse/systemtest/iot/x509/TrustAnchorTest.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.systemtest.iot.isolated.x509;
+package io.enmasse.systemtest.iot.x509;
 
 import io.enmasse.iot.model.v1.IoTProject;
 import io.enmasse.iot.model.v1.IoTProjectStatus;
